requirements.txt
python-telegram-bot==20.7
paramiko==3.4.0
asyncssh==2.14.2
aiofiles==23.2.1
aiohttp==3.9.1
cryptography==41.0.7
bcrypt==4.0.1
pynacl==1.5.0
psutil==5.9.6
python-dotenv==1.0.0
requests==2.31.0
typing-extensions==4.9.0

# pip install python-telegram-bot==20.7 paramiko==3.4.0 asyncssh==2.14.2 aiohttp==3.9.1 cryptography==41.0.7 psutil==5.9.6